requirements.txt
Flask==2.3.3
pandas==1.5.3
numpy==1.24.3
google-generativeai==0.3.1
Werkzeug==2.3.7
gunicorn==21.2.0
openai==1.3.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
